src/test/java/com/it/net/eureka/utils/CryptoUtilTest.java
package com.it.net.eureka.utils;

import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.junit.jupiter.api.Assertions.assertTrue;

import java.lang.reflect.Constructor;
import java.lang.reflect.InvocationTargetException;
import java.lang.reflect.Modifier;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.junit.jupiter.MockitoExtension;

@ExtendWith(MockitoExtension.class)
public class CryptoUtilTest {

	@BeforeEach
	protected void setUp() throws Exception {}

	@Test
	public final void testGenerateHashWithGivenSalt() throws Exception {
		assertNotNull(CryptoUtil.generateHashWithGivenSalt(Costants.STR_TEST, CryptoUtil.generateSalt()));
	}

	@Test
	public final void testPrivateConstructor() throws InstantiationException, IllegalAccessException, IllegalArgumentException, InvocationTargetException, NoSuchMethodException, SecurityException {
		Constructor<CryptoUtil> constructor = CryptoUtil.class.getDeclaredConstructor();
		assertTrue(Modifier.isPrivate(constructor.getModifiers()));
		constructor.setAccessible(true);
		assertThrows(InvocationTargetException.class, () -> constructor.newInstance());
	}
}	